+++ Templates/src/Core/PainKiller.PowerCommands.Configuration/DOMAINOBJECTS/ConfigurationGlobals.cs
@@ -8,7 +8,8 @@ public static class ConfigurationGlobals
     public const string WhatsNewFileName = "whats_new.md";
     public const char ArraySplitter = '|';
     public const string SetupConfigurationFile = "setup.yaml";
-    public const string EncryptionEnviromentVariableName = "_encryptionManager";
+    public const string EncryptionEnvironmentVariableName = "_encryptionManager";
 
     public static readonly string ApplicationDataFolder = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}\\{nameof(PowerCommands)}";
+    public static readonly string MainConfigurationFileFullPath = Path.Combine(AppContext.BaseDirectory, MainConfigurationFile);
 }