+++ src/PainKiller.PowerCommands/Core/PainKiller.PowerCommands.Core/Services/DialogService.cs
@@ -1,5 +1,4 @@
-﻿using PainKiller.PowerCommands.ReadLine;
-using PainKiller.PowerCommands.Security.Services;
+﻿using PainKiller.PowerCommands.Security.Services;
 
 namespace PainKiller.PowerCommands.Core.Services;
 public static class DialogService
@@ -39,63 +38,6 @@ public static class DialogService
         }
 
         return $"{secret}".Trim();
-    }
-    public static Dictionary<int, string> ListDialog(string header, List<string> items, bool multiSelect = false, bool autoSelectIfOnlyOneItem = true, ConsoleColor foregroundColor = ConsoleColor.White, ConsoleColor backgroundColor = ConsoleColor.Blue)
-    {
-        if (items.Count == 0) return new Dictionary<int, string>();
-        Console.Clear();
-        WriteHeader($"{header}\n");
-        var startRow = Console.CursorTop;
-        var startForegroundColor = Console.ForegroundColor;
-        var startBackgroundColor = Console.BackgroundColor;
-
-        for (int index=0; index<items.Count;index++)
-        {
-            var item = items[index];
-            Console.WriteLine($"{index+1}. {item}");
-        }
-        var quit = " ";
-        var input = "";
-        var selectedItems = new Dictionary<int, string>();
-
-        if (items.Count == 1 && autoSelectIfOnlyOneItem)
-        {
-            selectedItems.Add(0, items.First());
-            Console.WriteLine("");
-            Console.Write(" ");
-            ConsoleService.Service.WriteRowWithColor(1, foregroundColor, backgroundColor, $" 1. {items.First()}");
-            return selectedItems;
-        }
-
-        var label = multiSelect ? "Enter number(s) and hit enter" : "Enter a number and hit enter";
-        while (input != quit)
-        {
-            Console.WriteLine("");
-            Console.Write($"{label}>");
-            input = ReadLineService.Service.Read();
-            if(input.Trim() == "") break;
-            
-            var selectedIndex = (int.TryParse(input, out var index) ? index : 1);
-            if(selectedIndex > items.Count || selectedIndex < 1) selectedIndex = 1;
-            var selectedItem = new { Index = selectedIndex, Value = items[selectedIndex - 1] };
-            var itemAdded = selectedItems.TryAdd(selectedItem.Index-1, selectedItem.Value);
-            
-            ConsoleService.Service.ClearRow(Console.CursorTop-1);
-            var top = Console.CursorTop - 2;
-            Console.CursorTop = Math.Clamp(top, 0, Console.LargestWindowHeight - 1);
-            Console.CursorLeft = 0;
-
-            if(itemAdded) ConsoleService.Service.WriteRowWithColor(startRow + selectedIndex - 1, foregroundColor, backgroundColor, $"{selectedIndex}. {selectedItem.Value}");
-            else
-            {
-                selectedItems.Remove(selectedItem.Index);
-                ConsoleService.Service.WriteRowWithColor(startRow + selectedIndex - 1, startForegroundColor, startBackgroundColor, $"{selectedIndex}. {selectedItem.Value}");
-            }
-
-            if (!multiSelect) break;
-        }
-        ToolbarService.ClearToolbar();
-        return selectedItems;
     }
     private static void WriteHeader(string text)
     {