+++ src/PainKiller.PowerCommands/PainKiller.PowerCommands.Bootstrap/whats_new.md
@@ -1,4 +1,10 @@
 ﻿# What is new?
+## Version 1.0.3.2
+**Released 2023-12-26**
+- `DialogService.ListDialog` has major improvements with paging and select all functionality and have been moved to a new service, named `ListService`.
+- `ProxyCommand` now handles suggestion overrides in configuration.
+- `ConsoleTableService.RenderTable` generic constraint new() removed since it is not needed.
+- Bookmarks in `CdCommand` now handles `%USERNAME%` as a placeholder for current user in paths. 
 ## Version 1.0.3.1
 **Released 2023-11-30**
 - Adjusted the Power Command to be run with a service account and use encryption, update the setup to reflect this.